backend/requirements.txt
# Kalongo Hotel - Backend Dependencies
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
django-filter>=23.5
openpyxl>=3.1
reportlab>=4.0
Pillow>=10.0
qrcode[pil]>=7.4
django-environ>=0.11
dj-database-url>=2.1
whitenoise>=6.6
gunicorn>=21.0
